+++ feat/wikiDreamCancel/src/commonMain/kotlin/io/github/sophon/dreamcancel/data/MoveMapper.kt
@@ -1,11 +1,10 @@
 package io.github.sophon.dreamcancel.data
 
-import io.github.sophon.core.util.add2dAliases
 import io.github.sophon.core.util.cleanHtml
-import io.github.sophon.core.util.orDash
+import io.github.sophon.core.util.create2dAliases
 import io.github.sophon.core.util.decodeHtmlEntities
 import io.github.sophon.core.util.normalize2dInputs
-import io.github.sophon.core.util.useForwardVariantOnly
+import io.github.sophon.core.util.orDash
 import io.github.sophon.core.wiki.domain.model.Character
 import io.github.sophon.core.wiki.domain.model.Move
 import io.github.sophon.dreamcancel.FEATURE_URL
@@ -35,10 +34,9 @@ internal fun MoveDto.toDomain(
     val normalizedInput = this.input
         .orDash()
         .decodeHtmlEntities()
-        .normalize2dInputs(minimizeClose = false)
-        .useForwardVariantOnly()
+        .normalize2dInputs()
         .lowercase()
-    val aliasList = normalizedInput.add2dAliases() + normalizedInput.addAliasesForMultipleButtons()
+    val aliasList = normalizedInput.create2dAliases(isPartial = true)
 
     val move = Move(
         charName = character.displayName,
@@ -70,13 +68,4 @@ internal fun MoveDto.toDomain(
 
 internal fun formMoveWikiUrl(gameId: String, charName: String, name: String?): String {
     return "${FEATURE_URL}/$gameId/${charName.createQueryName()}/Data#${name.orEmpty().replace(" ", "_")}"
-}
-
-private fun String.addAliasesForMultipleButtons(): List<String> {
-    if (contains("/").not()) return listOf(this)
-
-    val base = substringBefore("/")
-    val buttons = substringAfter("/").map { it.toString() }
-
-    return listOf(base) + buttons.map { base.dropLast(1) + it }
 }